+++ Source/VerveAdapter.swift
@@ -68,13 +68,14 @@ final class VerveAdapter: PartnerAdapter {
     /// - parameter request: Information about the ad load request.
     /// - parameter completion: Closure to be performed with the fetched info.
     func fetchBidderInformation(request: PreBidRequest, completion: @escaping ([String : String]?) -> Void) {
-        guard let appToken else {
-            let error = error(.prebidFailureInvalidArgument, description: "App token is empty")
+        let signalData = HyBid.getCustomRequestSignalData("cb")
+        guard let signalData else {
+            let error = error(.prebidFailureInvalidArgument, description: "Signal data is empty")
             log(.fetchBidderInfoFailed(request, error: error))
             completion(nil)
             return
         }
-        completion(["app_auth_token": appToken])
+        completion(["signal_data": signalData])
     }
     
     /// Indicates if GDPR applies or not and the user's GDPR consent status.